requirements.txt
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-elevenlabs
python-dotenv